+++ lib/rstat/core_ext/fixnum/binomial_coefficient.rb
@@ -1,5 +1,5 @@
 class Fixnum
   def binomial_coefficient(bottom)
-    (self - bottom + 1 .. self).inject(1, &:*) / (2 .. bottom).inject(1, &:*)
+    (self - bottom + 1 .. self).inject(:*) / (2 .. bottom).inject(:*)
   end
 end